Add tests with jest
I needed to rename the namepace because the compilation creates a
variable with the first segment of the namespace. This was shadowing the
variable `test` from jest and the test was not being run.
```
var test=$CLJS.test || ($CLJS.test = {}); // before the change of namespace
var example=$CLJS.example || ($CLJS.example = {}); // after the change of namespace
```

Readme updated with instructions to run tests. There is an example
snapshot test.